refactor: split learn analyzer entrypoint
